gradient/requirements.txt
asgiref==3.6.0
attrs==18.2.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
click-completion==0.5.2
click-didyoumean==0.3.0
click-help-colors==0.9.1
colorama==0.4.3
Django==3.2.18
djangorestframework==3.14.0
drf-extra-fields==3.4.1
gql==3.0.0a6
gradient==2.0.6
gradient-utils==0.5.0
graphql-core==3.1.7
halo==0.0.31
idna==3.4
importlib-metadata==6.6.0
Jinja2==3.1.2
log-symbols==0.0.14
MarkupSafe==2.1.2
marshmallow==2.21.0
multidict==6.0.4
Pillow==9.5.0
progressbar2==4.2.0
prometheus-client==0.9.0
pymongo==3.13.0
python-dateutil==2.8.2
python-utils==3.5.2
pytz==2023.3
PyYAML==5.4.1
requests==2.30.0
requests-toolbelt==1.0.0
shellingham==1.5.0.post1
six==1.16.0
spinners==0.0.24
sqlparse==0.4.4
termcolor==2.3.0
terminaltables==3.1.10
typing_extensions==4.5.0
urllib3==2.0.2
websocket-client==0.57.0
windows-curses==2.3.1
yarl==1.9.2
zipp==3.15.0
django-cors-headers==4.0.0
django-embed-video==1.4.8